pyproject.toml
[project]
name = "att-site"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.4",
    "django-allauth[socialaccount]>=65.11.0",
    "django-environ>=0.12.0",
    "django-widget-tweaks>=1.5.0",
    "dotenv>=0.9.9",
    "environ>=1.0",
    "openpyxl>=3.1.5",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.1",
    "rapidfuzz>=3.13.0",
    "xlrd>=2.0.2",
]

[dependency-groups]
dev = []
